requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pymongo==4.7.3
python-dotenv==1.0.1
langchain==0.3.7
langchain-groq==0.2.1
langchain-core==0.3.15
dnspython==2.6.1